packaging alertmanager sets and add a tests.
